robot/requirements.txt
python-socketio==5.11.2
aiohttp==3.9.5
rich==13.7.1
PyYAML==6.0.1
